Final commit for about 12 hours, need sleep. Started work on Quartz Chest, it's Gui is about 85% done and working, just need to figure out how to modify the container from the gui and then it is complete. Chest is 90% done overall, but requires you to close and reopen the gui to update the tab contents.

+++ xan/storagecraft/tileentity/TileEntitySC.java
@@ -231,7 +231,7 @@ public class TileEntitySC extends TileEntity implements IInventory{
 			if (stack != null){
 				NBTTagCompound item = new NBTTagCompound();
 				
-				item.setByte("Slot", (byte)i);
+				item.setInteger("Slot", i);
 				stack.writeToNBT(item);
 				items.appendTag(item);
 			}
@@ -246,7 +246,7 @@ public class TileEntitySC extends TileEntity implements IInventory{
 		
 		for (int i = 0; i < items.tagCount(); i++){
 			NBTTagCompound item = (NBTTagCompound)items.tagAt(i);
-			int slot = item.getByte("Slot");
+			int slot = item.getInteger("Slot");
 			
 			if(slot >= 0 && slot < getSizeInventory()){
 				setInventorySlotContents(slot, ItemStack.loadItemStackFromNBT(item));
